requirements.txt
asgiref==3.7.2
certifi==2023.11.17
distlib==0.3.8
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
filelock==3.13.1
pipenv==2023.11.15
platformdirs==4.1.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.9.0
virtualenv==20.25.0
